NN-1796 Update add appointment form
