+++ app/components/Bookings/Details/AddAppointment/index.js
@@ -1,0 +1,3 @@
+import AddAppointment from './AddAppointment'
+
+export default AddAppointment
