+++ app/components/FormComponents/DatePicker/index.js
@@ -1,12 +1,23 @@
 import React, { Component } from 'react'
+import PropTypes from 'prop-types'
 import Datetime from 'react-datetime'
 import moment from 'moment'
 import 'react-datetime/css/react-datetime.css'
-import PropTypes from 'prop-types'
+import styled from 'styled-components'
+import LabelText from '@govuk-react/label-text'
+import Label from '@govuk-react/label'
+import ErrorText from '@govuk-react/error-text'
+import Input from '@govuk-react/input'
+import { MEDIA_QUERIES } from '@govuk-react/constants'
 
 import { DEFAULT_MOMENT_DATE_FORMAT_SPEC } from '../../../containers/App/constants'
-import './index.scss'
 import { inputType, metaType } from '../../../types'
+
+const StyledDatePicker = styled.div`
+  ${MEDIA_QUERIES.TABLET} {
+    max-width: 155px;
+  }
+`
 
 export class DatePicker extends Component {
   handleChange = date => {
@@ -14,46 +25,46 @@ export class DatePicker extends Component {
     input.onChange(date.format(DEFAULT_MOMENT_DATE_FORMAT_SPEC))
   }
 
-  renderInput = props => (
-    <div>
-      <input {...props} readOnly />
-    </div>
-  )
-
-  render() {
+  renderInput = props => {
     const {
       title,
-      defaultValue,
-      shouldShowDay,
-      locale,
-      showError,
-      input: { name, value },
+      input: { name },
       meta: { touched, error },
     } = this.props
 
     return (
+      <StyledDatePicker>
+        <Label error={touched && error} mb={6}>
+          <LabelText>{title}</LabelText>
+          {touched && error && <ErrorText>{error}</ErrorText>}
+          <Input {...props} name={name} error={touched && error} readOnly />
+        </Label>
+      </StyledDatePicker>
+    )
+  }
+
+  render() {
+    const {
+      defaultValue,
+      shouldShowDay,
+      locale,
+      input: { name, value },
+    } = this.props
+
+    return (
       <div className="date-picker-component">
-        <div className={((showError || (touched && error)) && 'form-group form-group-error') || 'form-group'}>
-          <label htmlFor={name} className="form-label">
-            {title}
-          </label>
-
-          <div className="error-message">{touched && error && <span>{error}</span>}</div>
-
-          <Datetime
-            className={(showError || (touched && error)) && 'form-control-error'}
-            onChange={this.handleChange}
-            timeFormat={false}
-            isValidDate={shouldShowDay}
-            locale={locale}
-            closeOnSelect
-            strictParsing
-            renderInput={this.renderInput}
-            defaultValue={defaultValue}
-            value={value}
-            inputProps={{ id: name }}
-          />
-        </div>
+        <Datetime
+          onChange={this.handleChange}
+          value={value}
+          timeFormat={false}
+          isValidDate={shouldShowDay}
+          locale={locale}
+          closeOnSelect
+          strictParsing
+          defaultValue={defaultValue}
+          renderInput={this.renderInput}
+          inputProps={{ id: name }}
+        />
       </div>
     )
   }
