--- app/components/FormComponents/DatePicker/index.scss
@@ -1,10 +1,0 @@
-.date-picker-component {
-
-  .form-group-error {
-    .rdtPicker {  /* stylelint-disable-line */
-      @include media-breakpoint-down(xs) {
-        margin-left: -1.5em !important;
-      }
-    }
-  }
-}
