--- app/containers/Bookings/Details/AddAppointment/index.scss
@@ -1,55 +1,0 @@
-.add-appointment {
-
-  h1{
-    margin-top: .5em;
-    margin-bottom: .5em;
-  }
-
-
-  @include media-breakpoint-down(xs) {
-    button:first-child {
-      @include add-gutter-margin-bottom();
-    }
-  }
-
-  .time-picker > .form-control {
-    width: 45% !important;
-    margin-right: 5px;
-  }
-
-  .rdtPicker {
-    width: 350px;
-
-    @include media-breakpoint-down(xs) {
-      width: 100%;
-    }
-  }
-
-  .col-xs-6 {
-    padding-right: 0 !important;
-  }
-
-  .form-control {
-    width: 100% !important;
-  }
-
-  select {
-    height: 2em;
-  }
-
-  textarea {
-    @include media-breakpoint-up(xs) {
-      height: 120px !important;
-    }
-  }
-
-  .shaded {
-    background: #E9F3F9;
-  }
-
-  .inset {
-    border-left: 2px solid rgb(191, 193, 195);
-    padding-left: 15px;
-    margin: 15px 0;
-  }
-}
