+++ app/containers/Bookings/Details/AddAppointmentContainer/index.js
@@ -1,0 +1,3 @@
+import AddAppointmentContainer from './AddAppointmentContainer'
+
+export default AddAppointmentContainer
