+++ package.json
@@ -53,15 +53,19 @@
     "@contentful/rich-text-types": "^13.1.0",
     "@govuk-react/checkbox": "^0.6.0",
     "@govuk-react/constants": "^0.6.0",
+    "@govuk-react/error-text": "^0.7.0",
     "@govuk-react/grid-col": "^0.7.0",
     "@govuk-react/grid-row": "^0.7.0",
     "@govuk-react/heading": "^0.6.0",
+    "@govuk-react/input": "^0.7.0",
     "@govuk-react/input-field": "^0.6.0",
+    "@govuk-react/label": "^0.7.0",
     "@govuk-react/label-text": "^0.6.0",
-    "@govuk-react/lib": "^0.6.0",
+    "@govuk-react/lib": "^0.7.0",
     "@govuk-react/radio": "^0.6.0",
-    "@govuk-react/select": "^0.6.0",
+    "@govuk-react/select": "^0.7.0",
     "@govuk-react/table": "^0.6.0",
+    "@govuk-react/text-area": "^0.7.0",
     "applicationinsights": "^1.0.0",
     "axios": "^0.18.0",
     "babel-core": "^7.0.0-bridge.0",
@@ -88,6 +92,7 @@
     "express": "^4.15.5",
     "express-interceptor": "^1.2.0",
     "extract-text-webpack-plugin": "^2.1.2",
+    "final-form": "^4.12.0",
     "fontfaceobserver": "2.0.7",
     "form-serialize": "^0.7.2",
     "get-stdin": "^6.0.0",
